Moved backup functions out of docker package.
